refactor(prize): modify package
LottoPrize 의 패키지 위치를 prize 로 변경했습니다.
