+++ src/main/java/lotto/domain/prize/LottoPrize.java
@@ -1,4 +1,4 @@
-package lotto.domain.matcher;
+package lotto.domain.prize;
 
 public enum LottoPrize {
